Same fixes for 32 bit Windows builds where std::size_t is a 32 bit integer; improved some comments
git-svn-id: file:///srv/devel/repo-conversion/nusu@230 d2e56fa2-650e-0410-a79f-9358c0239efd

+++ Source/Threading/Thread.cpp
@@ -76,7 +76,7 @@ namespace {
       // assumes that the thread is running on group 0.
       DWORD cpuCount = ::GetActiveProcessorCount(0);
       for(std::size_t index = 0; index < cpuCount; ++index) {
-        allCpusAffinity |= (std::size_t(1) << index);
+        allCpusAffinity |= (std::uint64_t(1) << index);
       }
     }
